requirements.txt
altair==4.2.0
asttokens==2.1.0
attrs==22.1.0
autopep8==2.0.0
backcall==0.2.0
blinker==1.5
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
confection==0.0.3
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
Cython==0.29.28
debugpy==1.6.3
decorator==5.1.1
docopt==0.6.2
emoji==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
entrypoints==0.4
executing==1.2.0
filelock==3.8.0
fonttools==4.38.0
gensim==4.2.0
gitdb==4.0.10
GitPython==3.1.29
huggingface-hub==0.11.0
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.1
jupyter-client==7.4.7
jupyter-core==5.0.0
keybert==0.7.0
keyphrase-vectorizers==0.0.10
kiwisolver==1.4.4
langcodes==3.3.0
language-tool-python==2.7.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
murmurhash==1.0.9
nest-asyncio==1.5.6
nltk==3.7
numpy==1.23.5
packaging==21.3
pandas==1.5.2
parso==0.8.3
pathy==0.10.0
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==2.5.4
preshed==3.0.8
prompt-toolkit==3.0.33
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==10.0.1
pycodestyle==2.10.0
pycontractions==2.0.2
pydantic==1.10.2
pydeck==0.8.0
pyemd==0.5.1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.3
scipy==1.9.3
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.4.3
spacy-alignments==0.8.6
spacy-legacy==3.0.10
spacy-loggers==1.0.3
spacy-transformers==1.1.8
srsly==2.4.5
stack-data==0.6.1
streamlit==1.15.1
thinc==8.1.5
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.21.3
typer==0.7.0
typing-extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
wasabi==0.10.1
watchdog==2.1.9
wcwidth==0.2.5
wget==3.2
wordcloud==1.8.2.2
yarg==0.1.9
zipp==3.10.0
